build-analyze.sh: Better modularity.

+++ tools/build-analyze.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
+MAKE="make -j1"
+OUTPUT=rainbow
 PROJECT="$(cd -P "$(dirname "${BASH_SOURCE[0]}")/.." && pwd)"
-REMOVABLE="CMake* cmake_install.cmake *.a Makefile rainbow"
+REMOVABLE="CMake* cmake_install.cmake *.a Makefile $OUTPUT"
 SCAN_BUILD=/usr/lib/clang-analyzer/scan-build
 
 if [ ! -d $SCAN_BUILD ]; then
@@ -8,19 +10,34 @@ if [ ! -d $SCAN_BUILD ]; then
 	exit 1
 fi
 
-# Test cross compiling
-rm -fr $REMOVABLE && \
-cmake -DMINGW=ON $PROJECT && \
-make -j1 && \
+function compile {
+	if ! which $2 &> /dev/null; then
+		echo $1 pass: Skipped
+		return
+	fi
+	rm -fr $REMOVABLE && \
+	CC=$CC CXX=$CXX cmake $3 $PROJECT && \
+	$MAKE
+	if [ ! -x $OUTPUT ]; then
+		echo $1 pass: Failed
+		exit 1
+	fi
+	echo $1 pass: OK
+}
 
-# Test compiling with GCC
-rm -fr $REMOVABLE && \
-CC=gcc CXX=g++ cmake $PROJECT && \
-make -j1 && \
+# Cross compile pass
+CC=
+CXX=
+compile "Cross compile" i486-mingw32-c++ -DMINGW=ON
 
-# Finally, compile with Clang analyzer and run tests
-rm -fr $REMOVABLE && \
-#CC=$SCAN_BUILD/ccc-analyzer CXX=$SCAN_BUILD/c++-analyzer cmake -DGTEST=ON $PROJECT && \
-CC=$SCAN_BUILD/ccc-analyzer CXX=$SCAN_BUILD/c++-analyzer cmake $PROJECT && \
-scan-build -o report --use-cc=clang --use-c++=clang make $@ && \
-./rainbow
+# GCC compile pass
+CC=gcc
+CXX=g++
+compile "GCC compile" $CC
+
+# Clang compile pass
+# Also performs static analysis and generates a report
+CC=$SCAN_BUILD/ccc-analyzer
+CXX=$SCAN_BUILD/c++-analyzer
+MAKE="scan-build -o report --use-cc=clang --use-c++=clang make $@"
+compile "Clang compile" scan-build #-DGTEST=ON
